SSCS-7578 Removing PIP-specific logic from base class
